getstring/filtertext cleanup
Former-commit-id: 1e72c194ecaf44569bf0270c6d7e58ab5e778072
Former-commit-id: 03fbe4037b18ad292409a936549fb3e9fab2451f

+++ src/engine/server.cpp
@@ -268,7 +268,7 @@ ICOMMAND(0, filterword, "ss", (char *s, char *t),
 });
 
 
-bool filtertext(char *dst, const char *src, bool newline, bool colour, bool whitespace, int len)
+bool filtertext(char *dst, const char *src, bool newline, bool colour, bool whitespace, size_t len)
 {
     bool filtered = false;
     for(int c = uchar(*src); c; c = uchar(*++src))
